Use latest version of aom and add setup scripts for rav1e and svt.
Given our project will be using the open source tools for av1, add setup
scripts for rav1e and svt so that we can use them for metric analysis.
The setup script for aom checked out a very old commit and was not to
decode the videos encoded by rav1e and svt. Thus, we updated the git
checkout to one of the recent versions and modified the setup commands
to reflect the current way of setting up aom