Add visual_metrics from aom and update code.
Since we will be using aom's tool to generate the html template,
Use the visual_metrics script from aom and add a helper function to
convert the results into stt files that the visual_metrics tool can take.

The other change made is that the graphs now will use 'actual-bitrate'
instead of 'target-bitrate'. This is due to the fact that we will also
have the option of playing around with qp_values. Therefore, we will not
always have target-bitrate